requirements.txt
albumentations
clip
einops
fairscale
huggingface_hub
imgaug
lmdb
loguru
matplotlib
mc.py
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
pandas
Pillow
pytest
PyYAML
recommonmark
Requests
rfconv
scikit_learn
scipy
setuptools
scikit-image
timm
torch==2.0.1
torchvision
tqdm
transformers
wandb
tensorboard